identification/requirements.txt
boto3==1.26.108
botocore==1.29.108
jmespath==1.0.1
joblib==1.2.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
python-dateutil==2.8.2
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15


torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101

